+++ CHANGELOG.md
@@ -5,16 +5,21 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [2.3.0] - 2024-08-01
-- Added new type - File - allows attachment and filename in same message.
+- Added new stype - file - allows attachment and filename in same message.
 - Updated rabbitmq-lib to v2.2.5
 - Updated simplejson==3.13.2
 - Updated requests==2.25.0
 - Added certifi==2019.11.28
 - Added idna==2.10
+- Added systemctl file for rmq-sysmon daemon.
 
 ### Changed
+- non_proc_msg: Added Routing Key information to email body.
+- run_program: Added date format to the log file name.
+- \_convert_data: Added check for stype of file for attachment and associated filename.
 - main: Removed parsing from gen_class.ArgParser call and called arg_parse2 as part of "if" statement.
 - Updates to requirements.txt.
+- Documentation updates.
 
 
 ## [2.2.2] - 2024-07-31